+++ .gitignore
@@ -12,6 +12,7 @@ security-patch.md
 frontend/next-env.d.ts
 PAPER_RAG_AGENT_PLAN.md
 ANONYMOUS_SESSION_AUTH_PLAN.md
+PERF.md
 
 # Next.js
 frontend/.next/
