+++ frontend/src/lib/layout-constants.ts
@@ -1,0 +1,10 @@
+// Timeline layout constants — the only live exports extracted from the old dummy-data module.
+const NODE_WIDTH = 260;
+const NODE_HEIGHT = 100;
+
+export const LANE_HEIGHT = 160;
+export const GAP_X = 80;
+export const PADDING_X = 60;
+export const PADDING_Y = 80;
+
+export const NODE_DIMENSIONS = { width: NODE_WIDTH, height: NODE_HEIGHT };
