+++ frontend/src/lib/timeline-builder.ts
@@ -1,4 +1,4 @@
-import { GAP_X, LANE_HEIGHT, NODE_DIMENSIONS, PADDING_X, PADDING_Y } from "./dummy-data";
+import { GAP_X, LANE_HEIGHT, NODE_DIMENSIONS, PADDING_X, PADDING_Y } from "./layout-constants";
 import { estimateTimelineNoteDimensions, layoutTimelineNotes } from "./note-layout";
 import { GraphEdge, GraphPaper, LineageGraphResponse, TimelineData, TimelineNode, TraceNote } from "./types";
